토큰 검증 어노테이션 이름 변경, DB, 로그인 기능 버그(유저 데이터를 못 불러옴) 해결

+++ src/main/java/gift/repository/WishRepository.java
@@ -1,6 +1,5 @@
 package gift.repository;
 
-import gift.dto.UserInfoRequestDto;
 import gift.entity.Wish;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
@@ -13,21 +12,21 @@ public class WishRepository {
     private final JdbcTemplate jdbcTemplate;
     public WishRepository(JdbcTemplate jdbcTemplate) {this.jdbcTemplate = jdbcTemplate;}
 
-    public List<Wish> findUserWishes(UserInfoRequestDto userInfoRequestDto) {
-        return jdbcTemplate.query("SELECT id, user_id, product_id, quantity FROM wishes WHERE user_id = ?", wishRowMapper(), userInfoRequestDto.id());
+    public List<Wish> findUserWishes(Long userId) {
+        return jdbcTemplate.query("SELECT id, user_id, product_id, quantity FROM wishes WHERE user_id = ?", wishRowMapper(), userId);
     }
 
-    public Wish addWish(Wish wish) {
-        jdbcTemplate.update("INSERT INTO wishes(id, user_id, product_id, quantity) VALUES(?, ?, ?)", wish.userId(), wish.productId(), wish.quantity());
+    public Wish addWish(Long userId, Wish wish) {
+        jdbcTemplate.update("INSERT INTO wishes(user_id, product_id, quantity) VALUES(?, ?, ?)", userId, wish.productId(), wish.quantity());
         return wish;
     }
 
-    public void updateWish(Wish wish) {
-        jdbcTemplate.update("UPDATE wishes SET quantity = ? WHERE id = ?", wish.quantity(), wish.id());
+    public void updateWish(Long userId, Wish wish) {
+        jdbcTemplate.update("UPDATE wishes SET quantity = ? WHERE id = ? AND user_id = ?", wish.quantity(), wish.id(), userId);
     }
 
-    public void deleteWish(Wish wish) {
-        jdbcTemplate.update("DELETE FROM wishes WHERE id = ?", wish.id());
+    public void deleteWish(Long userId, Wish wish) {
+        jdbcTemplate.update("DELETE FROM wishes WHERE id = ? AND user_id = ?", wish.id(), userId);
     }
 
     // Wishes 조회용 RowMapper
